pyproject.toml
[project]
name = "kbtransfer"
version = "0.1.0"
description = "Open-protocol federated knowledge base between AI agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "KBTRANSFER contributors" },
]
keywords = ["mcp", "agents", "knowledge-base", "karpathy-wiki", "autoevolve-packs"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "cryptography>=42.0",
    "pyyaml>=6.0",
    "mcp>=1.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.8",
    "httpx>=0.27",
    "fastapi>=0.110",
]
server = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
]

[project.scripts]
kb = "kb_cli.__main__:main"
kb-mcp = "kb_mcp_server.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "reference/kb_cli",
    "reference/kb_mcp_server",
    "reference/kb_pack",
    "reference/kb_distiller",
    "reference/kb_registry",
    "reference/kb_registry_server",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
